Separated editor gestures/key recognizers from EditableDocument, which brings all the core pieces together: document, composer, interactor.
